Move encryption/tls related functions to encryption package (#199)
* Move encryption/tls related functions to encryption package

* Export cert stucts
