+++ pkg/encryption/rotation.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package encryption
 
 import (
 	"crypto/tls"
@@ -23,10 +23,11 @@ import (
 
 	"github.com/fsnotify/fsnotify"
 	"github.com/gogatekeeper/gatekeeper/pkg/utils"
+	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 )
 
-type certificationRotation struct {
+type CertificationRotation struct {
 	sync.RWMutex
 	// certificate holds the current issuing certificate
 	certificate tls.Certificate
@@ -35,11 +36,12 @@ type certificationRotation struct {
 	// the privateKeyFile is the path of the private key
 	privateKeyFile string
 	// the logger for this service
-	log *zap.Logger
+	log            *zap.Logger
+	rotationMetric *prometheus.Counter
 }
 
 // newCertificateRotator creates a new certificate
-func newCertificateRotator(cert, key string, log *zap.Logger) (*certificationRotation, error) {
+func NewCertificateRotator(cert, key string, log *zap.Logger, metric *prometheus.Counter) (*CertificationRotation, error) {
 	// step: attempt to load the certificate
 	certificate, err := tls.LoadX509KeyPair(cert, key)
 
@@ -48,16 +50,17 @@ func newCertificateRotator(cert, key string, log *zap.Logger) (*certificationRot
 	}
 
 	// @step: are we watching the files for changes?
-	return &certificationRotation{
+	return &CertificationRotation{
 		certificate:     certificate,
 		certificateFile: cert,
 		log:             log,
 		privateKeyFile:  key,
+		rotationMetric:  metric,
 	}, nil
 }
 
 // watch is responsible for adding a file notification and watch on the files for changes
-func (c *certificationRotation) watch() error {
+func (c *CertificationRotation) Watch() error {
 	c.log.Info(
 		"adding a file watch on the certificates, certificate",
 		zap.String("certificate", c.certificateFile),
@@ -97,7 +100,7 @@ func (c *certificationRotation) watch() error {
 							zap.Error(err))
 					}
 					// @metric inform of the rotation
-					certificateRotationMetric.Inc()
+					(*c.rotationMetric).Inc()
 					// step: load the new certificate
 					_ = c.storeCertificate(certificate)
 					// step: print a debug message for us
@@ -113,7 +116,7 @@ func (c *certificationRotation) watch() error {
 }
 
 // storeCertificate provides entrypoint to update the certificate
-func (c *certificationRotation) storeCertificate(certifacte tls.Certificate) error {
+func (c *CertificationRotation) storeCertificate(certifacte tls.Certificate) error {
 	c.Lock()
 	defer c.Unlock()
 	c.certificate = certifacte
@@ -122,7 +125,7 @@ func (c *certificationRotation) storeCertificate(certifacte tls.Certificate) err
 }
 
 // GetCertificate is responsible for retrieving
-func (c *certificationRotation) GetCertificate(hello *tls.ClientHelloInfo) (*tls.Certificate, error) {
+func (c *CertificationRotation) GetCertificate(hello *tls.ClientHelloInfo) (*tls.Certificate, error) {
 	c.RLock()
 	defer c.RUnlock()
