+++ pkg/utils/utils_test.go
@@ -19,7 +19,6 @@ limitations under the License.
 package utils
 
 import (
-	"bytes"
 	"crypto/tls"
 	"fmt"
 	"io/ioutil"
@@ -33,7 +32,6 @@ import (
 	uuid "github.com/gofrs/uuid"
 	"github.com/gogatekeeper/gatekeeper/pkg/constant"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestDecodeKeyPairs(t *testing.T) {
@@ -180,50 +178,6 @@ func TestDefaultTo(t *testing.T) {
 	}
 }
 
-func TestEncryptDataBlock(t *testing.T) {
-	testCase := []struct {
-		Text string
-		Key  string
-		Ok   bool
-	}{
-		{
-			Text: "hello world, my name is Gatekeeper",
-			Key:  "DtNMS2eO7Fi5vsuLrW55nrRbir2kPfTw",
-			Ok:   true,
-		},
-		{
-			Text: "hello world, my name is Gatekeeper",
-			Key:  "DtNMS2eO7Fi5vsu",
-		},
-		{
-			Text: "h",
-			Key:  "DtNMS2eO7Fi5vsuLrW55nrRbir2kPfTwtr",
-		},
-	}
-
-	for i, test := range testCase {
-		_, err := EncryptDataBlock(bytes.NewBufferString(test.Text).Bytes(), bytes.NewBufferString(test.Key).Bytes())
-		if err != nil && test.Ok {
-			t.Errorf("test case: %d should not have failed, %s", i, err)
-		}
-	}
-}
-
-func TestEncodeText(t *testing.T) {
-	session, err := EncodeText("12245325632323263762", "1gjrlcjQ8RyKANngp9607txr5fF5fhf1")
-	assert.NotEmpty(t, session)
-	assert.NoError(t, err)
-}
-
-var (
-	fakePlainText = []byte(`nFlhnhwRzC9uJ9mjhR0PQezUpIiDlU9ASLqH1KIKFhBZZrMZfnAAdHdgKs2OJoni8cTSQ
-	JxkaNpboZ6hnrMytlw5kf0biF7dLTU885uHIGkUIRy75hx6BaTEEhbN36qVTxediEHd6xeBPS3qpJ7riO6J
-	EeaQr1rroDL0LvmDyB6Zds4LdVQEmtUueusc7jkBz7gJ12vnTHIxviZM5rzcq4tyCbZO7Kb37RqZg5kbYGK
-	PfErhUwUIin7jsNVE7coB`)
-	fakeCipherText = []byte("lfQPTa6jwMTABaJhcrfVkoqcdyMVAettMsqgKXIALSKG5UpoYKbT/WgZjOiuCmEI0E/7piP8VATLOAHKDBNF2WrQOKSYF+gdHkh4NLv0cW0NZ2qyZeWhknywE6063ylhCYjJOrJA1z12i2bHHbjZZGfqkwfzyxxFLTv6jSbalpZ4oZcUcNY/DrtVk/K01qZw6o4l1f0FUL6UZVSirn+B3YDWLeVQ0FGr6jlhCpN203Rf688nqdBvhw4bUEQiykCMxWm2/rJBNWm2SzZgw65kb4W0ph1qjcoUjXBwNakK+E0Lw/fwi8+bUC1lkT8+hJpMLKZkzb07rbGAnmljQo0NkqJh4kl+aycsEhm9bZj+b6w0r795YugyNsyca5CnUvkB1Dg")
-	fakeKey        = []byte("u3K0eKsmGl76jY1buzexwYoRRLLQrQck")
-)
-
 /*
 func TestEncryptedText(t *testing.T) {
 	s, err := encodeText(string(fakePlainText), string(fakeKey))
@@ -238,90 +192,12 @@ func TestEncryptedText(t *testing.T) {
 }
 */
 
-func BenchmarkEncryptDataBlock(b *testing.B) {
-	for n := 0; n < b.N; n++ {
-		_, _ = EncryptDataBlock(fakePlainText, fakeKey)
-	}
-}
-
-func BenchmarkEncodeText(b *testing.B) {
-	text := string(fakePlainText)
-	key := string(fakeKey)
-	for n := 0; n < b.N; n++ {
-		_, _ = EncodeText(text, key)
-	}
-}
-
-func BenchmarkDecodeText(b *testing.B) {
-	t := string(fakeCipherText)
-	k := string(fakeKey)
-	for n := 0; n < b.N; n++ {
-		if _, err := DecodeText(t, k); err != nil {
-			b.FailNow()
-		}
-	}
-}
-
-func TestDecodeText(t *testing.T) {
-	fakeKey := "HYLNt2JSzD7Lpz0djTRudmlOpbwx1oHB"
-	fakeText := "12245325632323263762"
-
-	encrypted, err := EncodeText(fakeText, fakeKey)
-	require.NoError(t, err)
-	assert.NotEmpty(t, encrypted)
-
-	decoded, _ := DecodeText(encrypted, fakeKey)
-	assert.NotNil(t, decoded, "the session should not have been nil")
-	assert.Equal(t, decoded, fakeText, "the decoded text is not the same")
-}
-
 func TestFindCookie(t *testing.T) {
 	cookies := []*http.Cookie{
 		{Name: "cookie_there"},
 	}
 	assert.NotNil(t, FindCookie("cookie_there", cookies))
 	assert.Nil(t, FindCookie("not_there", cookies))
-}
-
-func TestDecryptDataBlock(t *testing.T) {
-	testCase := []struct {
-		Text string
-		Key  string
-		Ok   bool
-	}{
-		{
-			Text: "hello world, my name is Gatekeeper",
-			Key:  "DtNMS2eO7Fi5vsuLrW55nrRbir2kPfss",
-			Ok:   true,
-		},
-		{
-			Text: "h",
-			Key:  "DtNMS2eO7Fi5vsuLrW55nrRbir2kPfTw",
-			Ok:   true,
-		},
-	}
-
-	for idx, test := range testCase {
-		cipher, err := EncryptDataBlock(
-			bytes.NewBufferString(test.Text).Bytes(),
-			bytes.NewBufferString(test.Key).Bytes(),
-		)
-		if err != nil && test.Ok {
-			t.Errorf("test case: %d should not have failed, %s", idx, err)
-		}
-
-		plain, err := DecryptDataBlock(
-			cipher,
-			bytes.NewBufferString(test.Key).Bytes(),
-		)
-		if err != nil {
-			t.Errorf("test case: %d should not have failed, %s", idx, err)
-		}
-
-		if string(plain) != test.Text {
-			t.Errorf("test case: %d are not the same", idx)
-		}
-	}
 }
 
 func TestHasAccessOK(t *testing.T) {
